fix(02): IN-05 widen Freshdesk id columns to BIGINT to avoid overflow
Real Freshdesk conversation/ticket ids already exceed 2^31 (observed
60293425282). Change ticket_id and inbound_msg_id from INTEGER to BIGINT
in queue.ticket_queue, queue.dead_letter, and queue.dry_run_log so inserts
of real ids do not raise numeric-overflow (breaks REP-05). Greenfield
migration 0001 edited in place; freshdesk_reply_id already BIGINT,
attempts/max_attempts stay INTEGER, poller_checkpoint.id CHECK(id=1) stays
INTEGER.

+++ migrations/versions/0001_initial_queue_schema.py
@@ -41,8 +41,8 @@ def upgrade() -> None:
         CREATE TABLE queue.ticket_queue (
             id                  BIGSERIAL PRIMARY KEY,
             idempotency_key     TEXT        NOT NULL,
-            ticket_id           INTEGER     NOT NULL,
-            inbound_msg_id      INTEGER     NOT NULL,
+            ticket_id           BIGINT      NOT NULL,
+            inbound_msg_id      BIGINT      NOT NULL,
             payload             JSONB       NOT NULL,
             status              TEXT        NOT NULL DEFAULT 'pending',
             claimed_at          TIMESTAMPTZ,
@@ -86,8 +86,8 @@ def upgrade() -> None:
         CREATE TABLE queue.dead_letter (
             id              BIGSERIAL   PRIMARY KEY,
             idempotency_key TEXT        NOT NULL,
-            ticket_id       INTEGER     NOT NULL,
-            inbound_msg_id  INTEGER     NOT NULL,
+            ticket_id       BIGINT      NOT NULL,
+            inbound_msg_id  BIGINT      NOT NULL,
             payload         JSONB       NOT NULL,
             attempts        INTEGER     NOT NULL,
             last_error      TEXT,
@@ -104,8 +104,8 @@ def upgrade() -> None:
         """
         CREATE TABLE queue.dry_run_log (
             id              BIGSERIAL   PRIMARY KEY,
-            ticket_id       INTEGER     NOT NULL,
-            inbound_msg_id  INTEGER     NOT NULL,
+            ticket_id       BIGINT      NOT NULL,
+            inbound_msg_id  BIGINT      NOT NULL,
             action          TEXT        NOT NULL,
             body            TEXT        NOT NULL,
             created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
